backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
numpy<2
opencv-python==4.10.0.84
torch==2.3.1
torchvision==0.18.1
ultralytics==8.3.0
Werkzeug==2.3.7
python-dotenv==1.0.0
twilio
